+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v2.0.6] - 2025-11-24
+
 ### Added
+- Add pantry Queue
 - Add handler for daily pantry stats
 
 ## [v2.0.5] - 2025-11-19